Switch to android-maven-plugin 3.7.0 again, fixing the crash on Android 2.3
Details: Compiling with android-maven-plugin 3.8.2 (which requires Maven 3.1.1+)
results in an ActionBarSherlock NPE during startup on pre-4.0 devices.

When compiling with android-maven-plugin 3.7.0 (which requires Maven 3.0.5 or
older) this problem does not occur. Unfortunately this version of the plugin
does not yet support incremental dexing, making the build slower. This patch
activates pre-dexing support instead, which also speeds up the build, but not as
much.